Spliting instructions from Code Generation
