+++ codegenerator.cc
@@ -8,6 +8,7 @@
 #include <fstream>
 #include "mifgenerator.h"
 #include "bytes.h"
+#include "instructions.h"
 
 ConditionCodes translateCondition(TokenType operation)
 {
@@ -684,112 +685,7 @@ void CodeGenerator::createFooter()
         std::cout << "This is a footer\n";
 }
 
-//Instruction Class
-std::string Instruction::to_string()
-{
-    return "Generic Instruction";
-}
 
-Instruction *jumpToRegister(Registers reg)
-{
-    return new TypeFInstruction(
-        38,
-        "BX",
-        AL,
-        reg
-
-    );
-}
-Instruction *outputRegister(Registers reg)
-{
-    return new TypeEInstruction(
-        69,
-        "OUTPUT",
-        0,
-        reg);
-}
-
-Instruction *
-popRegister(Registers reg)
-{
-    return new TypeEInstruction(
-        68,
-        "POP",
-        0,
-        reg);
-}
-
-Instruction *pushRegister(Registers reg)
-{
-    return new TypeEInstruction(
-        67,
-        "PUSH",
-        0,
-        reg);
-}
-
-Instruction *loadImediateToRegister(Registers regis, int number)
-{
-    return new TypeDInstruction(
-        8,
-        "MOV",
-        regis,
-        number);
-}
-
-Instruction *pushAcumulator()
-{
-    return pushRegister(AcumulatorRegister);
-}
-
-Instruction *nop()
-{
-    return new TypeDInstruction(74, "NOP", 0, 0);
-}
-
-Instruction *moveLowToLowRegister(Registers origin, Registers destination)
-{
-    return new TypeCInstruction(
-        6,
-        "ADD",
-        0,
-        origin,
-        destination);
-}
-
-Instruction *subImeditateFromRegister(int value, Registers destination)
-{
-    // return new TypeDInstruction(
-    //     11,
-    //     "SUB",
-    //     destination,
-    //     value
-    // );
-    return new TypeCInstruction(
-        7,
-        "SUB",
-        value,
-        destination,
-        destination);
-}
-
-Instruction *moveLowToHigh(Registers low, Registers high)
-{
-    return new TypeEInstruction(
-        36,
-        "MOV",
-        low,
-        high);
-}
-
-Instruction *moveHighToLow(Registers low, Registers high)
-{
-    return new TypeEInstruction(
-        35,
-        "MOV",
-        high,
-        low);
-}
 
 void CodeGenerator::loadVariable(TreeNode *node, Registers reg)
 {
@@ -1080,26 +976,3 @@ void CodeGenerator::generateCodeForConst(int value)
     }
 }
 
-std::string printRegister(int reg)
-{
-    switch (reg)
-    {
-    case HeapArrayRegister:
-        return "$H0";
-    case AcumulatorRegister:
-        return "$A0";
-    case TemporaryRegister:
-        return "$T1";
-    case FramePointer:
-        return "$FP";
-    case GlobalPointer:
-        return "$GP";
-    case BaseAddressRegister:
-        return "$BA";
-    case ReturnAddressRegister:
-        return "$RA";
-    case SwapRegister:
-        return "$SR";
-    }
-    return "!UNKNOWN!";
-}
